refactor(sync): extract gamedata_pair state machine (py)
P2.B.2 (Python half). Move the GameData-pair state machine
(sync_release_archive / sync_release_archive_pair + pair path/load/save/init)
to sync/gamedata_pair, carrying BOTH #152/#155 idempotency guards verbatim
(archive-layer "already at activation_sha" short-circuit + pair-layer
no-op-write short-circuit). data/sync.py is now a pure re-export barrel.

Repoint the test patches whose caller moved (caller-timing principle):
sync_release (12) and sync_release_archive (4) -> sync.gamedata_pair. The
cross-process subprocess test is rewritten to patch
prts_mcp.sync.gamedata_pair.sync_release (where sync_release_archive now
resolves it) with a sentinel the parent asserts, so silent non-interception
fails loud. The #152 regression test is unchanged and green.

Folds the stranded P2.B.1 nit: sync/__init__.py docstring now reflects that
the full state machine lives in sync/. 369 passed / 23 skipped.
