+++ python/src/prts_mcp/sync/__init__.py
@@ -1,6 +1,7 @@
 """GitHub Release data-sync tier.
 
-The only layer permitted to issue HTTP. Holds the transport (HTTP / mirror
-cascade) and release-discovery leaves extracted from ``data/sync``; the
-release/archive/pair state machine still lives in ``data/sync`` until P2.B.
+The only layer permitted to issue HTTP. Holds the full sync state machine:
+transport (HTTP / mirror cascade), release discovery, release-archive
+activation, the release state machine, and the GameData-pair state machine.
+``data/sync`` is now a pure re-export barrel.
 """
